// api/generated/models/exerciseResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lingostan API
 * OpenAPI spec version: 1.0
 */
import type { ExerciseResponseDtoContent } from './exerciseResponseDtoContent';

export interface ExerciseResponseDto {
  id: number;
  type: string;
  title: string;
  instructions: string;
  content: ExerciseResponseDtoContent;
  points: number;
  order: number;
  hints: string[];
  explanation: string;
  isActive: boolean;
  lessonId: number;
  completed: boolean;
  userScore: number;
  attempts: number;
  createdAt: string;
}